Move standalone code to a new file
