+++ index.js
@@ -1,42 +1,29 @@
-/* eslint-disable no-console */
-import express from 'express'
 import { ApolloServer } from 'apollo-server-express'
 import { typeDefs } from './typeDefs'
 import { resolvers } from './resolvers'
 import mongoose from 'mongoose'
 import cachegoose from 'cachegoose'
 
-const PORT = 4000
-const DB_NAME = 'courseapi'
-const MONGO_URI = `mongodb+srv://test-a:nebku0-hYpqeq-qagmuh@cluster0-ydk8h.mongodb.net/${DB_NAME}?retryWrites=true`
+// DEBUG
+mongoose.set('debug', true)
 
-function connectMongo() {
-  // DEBUG
-  mongoose.set('debug', true)
-  mongoose.connect(MONGO_URI, { useNewUrlParser: true })
-
+function connectMongo(mongoURI) {
+  mongoose.connect(mongoURI, { useNewUrlParser: true })
   cachegoose(mongoose)
 }
 
-const server = new ApolloServer({
-  typeDefs,
-  resolvers,
-  context: async () => ({
-    mongo: await connectMongo()
-  }),
-  cacheControl: {
-    defaultMaxAge: 240
-  },
-  introspection: true,
-  playground: true,
-})
-
-const app = express()
-
-server.applyMiddleware({ app }) // app is from an existing express app
-
-app.listen({ port: PORT }, () =>
-  console.log(
-    `🚀 Server ready at http://localhost:${PORT}${server.graphqlPath}`
-  )
-)
+// CourseApiServer returns an ApolloServer connected to the MongoDB database
+export function CourseApiServer(mongoURI) {
+  return new ApolloServer({
+    typeDefs,
+    resolvers,
+    context: async () => ({
+      mongo: await connectMongo(mongoURI)
+    }),
+    cacheControl: {
+      defaultMaxAge: 240
+    },
+    introspection: true,
+    playground: true,
+  })
+}
